+++ CHANGELOG.md
@@ -10,7 +10,11 @@ The format is based on the [KeepAChangeLog] project.
 ### Changed
 - [#810] Drop python 3.6 support
 
+### Fixed
+- [#812] Fixed parsing of zero content lenght responses
+
 [#810]: https://github.com/OpenIDC/pyoidc/pull/810
+[#812]: https://github.com/OpenIDC/pyoidc/pull/812
 
 ## 1.3.0 [2021-07-09]